+++ wifeleviro/src/br/com/wifeleviro/ad/modelo/Mensagem.java
@@ -10,8 +10,8 @@ import br.com.wifeleviro.ad.util.GeradorRandomicoSingleton;
  */
 public class Mensagem {
 
-	public static final double TEMPO_TRANSMISAO_POR_QUADRO = 0.0008; //segundos por quadro 
-	public static final double TEMPO_TRANSMISSAO_REFORCO_COLISAO = 0.0000032; //segundos por reforco
+	public static final double TEMPO_TRANSMISAO_POR_QUADRO = (double)0.0008; //segundos por quadro 
+	public static final double TEMPO_TRANSMISSAO_REFORCO_COLISAO = (double)0.0000032; //segundos por reforco
 	
 	public static final int MENSAGEM_PADRAO = 0;
 	public static final int REFORCO_COLISAO = 1;
@@ -19,7 +19,7 @@ public class Mensagem {
 	private long id;
 	private int rodada;
 	
-	private double numeroQuadros;
+	private long numeroQuadros;
 	private int numeroQuadroRestantesParaTransmissao;
 	private int tipoMensagem;
 	
@@ -40,14 +40,14 @@ public class Mensagem {
 		if(p > 0 && p < 1){
 			double q = 1 - p;
 			Random r = new Random();
-			this.numeroQuadros = Math.log(r.nextDouble()%1)/Math.log(q);
+			this.numeroQuadros = (long)(Math.log(r.nextDouble()%1)/Math.log(q));
 		}else{
-			this.numeroQuadros = p;
+			this.numeroQuadros = (long)p;
 		}
 		this.numeroQuadroRestantesParaTransmissao = (int)this.numeroQuadros;
 	}
 	
-	public double getNumeroQuadros() {
+	public long getNumeroQuadros() {
 		return numeroQuadros;
 	}
